Time: 0 ms (100.00%) | Memory: 42 MB (64.00%) - LeetSync

+++ 50-powx-n/powx-n.java
@@ -14,7 +14,7 @@ class Solution {
                 ans = ans * x;
             }
         }
-        if(nn < 0) return (double)1.0/(double) ans;
+        if(nn < 0) return 1.0/ ans;
         return ans;
     }
 }
